+++ django_twilio_access_token/views.py
@@ -3,7 +3,7 @@ from rest_framework import status, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
-from .models import TwilioAccessToken
+from .tokens import TwilioAccessToken
 from .serializers import VideoTokenDeserializer
